requirements.txt
streamlit>=1.28.0
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
plotly>=5.14.1
python-dateutil>=2.8.2
